Use bigint for the score

+++ migrations/2023-07-29-170058_highscore/up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE highscore (
     id INTEGER PRIMARY KEY NOT NULL, -- an INTEGER PRIMARY KEY will auto-increment in Sqlite.
     name VARCHAR(255) NOT NULL,
-    score INTEGER NOT NULL,
+    score BIGINT NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL
 )
